refactor: remove the old app sidebar
